Time: 4 ms (82.41%) | Memory: 8 MB (18.31%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        long int z=x;
+        int z=x;
         long int y=0;
         if(x<0){
             return false;
